Edit calculator functions

+++ calculator.cpp
@@ -2,9 +2,9 @@
 #include <iostream>
 #include <cstdlib>
 
-int add(int a, int b) { return a + b; }
-int subtract(int a, int b) { return a - b; }
-int multiply(int a, int b) { return a * b; }
+double add(double a, double b) { return a + b; }
+double subtract(double a, double b) { return a - b; }
+double multiply(double a, double b) { return a * b; }
 double divide(double a, double b) { return (b != 0) ? a / b : 0; }
 
 long long factorial(int n) {
